requirements.txt
albumentations
boltons
cytoolz
Jinja2
lmdb
matplotlib
numpy
pandas
Pillow
seaborn
sklearn
torch
torchvision
tqdm
